requirements.txt
discord.py
aiohttp
python-whois
phonenumbers
dnspython
python-dotenv
requests